Archiving app directory as a legacy directory: renamed auth.py to auth_legacy.py
