+++ app_legacy/test.py
@@ -1,5 +1,6 @@
 from pymongo import MongoClient
 from pymongo.errors import ConnectionFailure
+from flask_pymongo import PyMongo
 
 # Replace with your Mongo URI
 uri = "mongodb+srv://nfldty:GAx4aJuymwNUKOgE@ploggo.qm3fx.mongodb.net/?retryWrites=true&w=majority&appName=PlogGo"
@@ -16,6 +17,7 @@ try:
     db = client.get_database("PlogGo")  # Replace with your database name
 
     print(f"Connected to the database: {db.name}")
+
     
 except ConnectionFailure as e:
     print(f"Could not connect to MongoDB: {e}")
